+++ front-end/.dockerignore
