+++ front-end/package.json
@@ -20,7 +20,8 @@
     "react-bootstrap": "^2.10.8",
     "react-dom": "^18.3.1",
     "react-query": "^3.39.3",
-    "react-router-dom": "^7.1.4"
+    "react-router-dom": "^7.1.4",
+    "source-map-support": "^0.5.21"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.1",
@@ -32,6 +33,6 @@
     "globals": "^15.9.0",
     "postcss": "^8.4.35",
     "tailwindcss": "^3.4.1",
-    "vite": "^5.4.2"
+    "vite": "^5.4.14"
   }
 }
