+++ front-end/src/components/ProjectManager.jsx
