+++ front-end/src/components/idea.md
