+++ front-end/src/components/stake.jsx
